backend/requirements.txt
fastapi==0.110.1
starlette==0.37.2
uvicorn==0.27.1
web3==6.8.0
eth-utils==2.3.0
eth-typing==3.6.0
python-dotenv==1.0.1
pytest==7.4.2
pytest-asyncio==0.23.5
httpx==0.27.0
